--- a/icicle/backend/cpu/src/curve/cpu_msm.hpp
+++ b/icicle/backend/cpu/src/curve/cpu_msm.hpp
@@ -732,10 +732,7 @@ template <typename A, typename P>
 eIcicleError cpu_msm(
   const Device& device, const scalar_t* scalars, const A* bases, int msm_size, const MSMConfig& config, P* results)
 {
-  int c = config.c;
-  if (c < 1) { c = Msm<A, P>::get_optimal_c(msm_size, config.precompute_factor); }
-
-  int nof_threads = std::thread::hardware_concurrency() - 1;
+  int nof_threads = std::thread::hardware_concurrency();
   if (config.ext && config.ext->has(CpuBackendConfig::CPU_NOF_THREADS)) {
     nof_threads = config.ext->get<int>(CpuBackendConfig::CPU_NOF_THREADS);
   }
@@ -743,7 +740,65 @@ eIcicleError cpu_msm(
     ICICLE_LOG_WARNING << "Unable to detect number of hardware supported threads - fixing it to 1\n";
     nof_threads = 1;
   }
-  auto msm = Msm<A, P>{config, c, nof_threads};
+  ICICLE_LOG_INFO << "nof_threads=" << nof_threads;
+
+  constexpr int nof_workers_per_msm = 8;
+  const int nof_top_level_msms =
+    (int)std::ceil((float)nof_threads / nof_workers_per_msm); // 4 workers per msm to avoid too much load
+  std::vector<P> sub_results(nof_top_level_msms, P::zero());  // Results for each thread
+  std::vector<std::thread> threads;
+
+  const int stride = (int)std::ceil((float)msm_size / nof_top_level_msms); // Workload per thread
+
+  ICICLE_LOG_INFO << "msm_size=" << msm_size << ", nof_workers_per_msm=" << nof_workers_per_msm
+                  << ", nof_top_level_msms=" << nof_top_level_msms << ", stride=" << stride;
+
+  // Launch threads
+  for (int i = 0; i < nof_top_level_msms; ++i) {
+    const bool last_msm = i == nof_top_level_msms - 1;
+    const int single_msm_size = last_msm ? (msm_size - (nof_top_level_msms - 1) * stride) : stride;
+
+    // ICICLE_LOG_INFO << msm_size << " - "
+    //                 << "(" << nof_top_level_msms << " - 1) * " << stride;
+    ICICLE_LOG_INFO << "i=" << i << ", last=" << last_msm << ", size=" << single_msm_size;
+
+    threads.emplace_back([&, i] { // Capture by reference (&) and index by value (i)
+      cpu_main(
+        scalars + i * stride, // Scalars for this thread
+        bases + i * stride,   // Bases for this thread
+        single_msm_size,      // Size for this thread
+        config,               // Config
+        &sub_results[i],      // Output to the sub-results array
+        nof_workers_per_msm /*for main*/);
+    });
+
+    // When not commented out, this is waiting for each sub-msm and proves the top level is correct
+    // threads.back().join(); // TODO remove
+    // threads.clear();
+  }
+
+  // Wait for threads
+  for (auto& thread : threads) {
+    if (thread.joinable()) { thread.join(); }
+  }
+
+  // Combine results
+  *results = sub_results[0];
+  for (int i = 1; i < sub_results.size(); ++i) {
+    *results = *results + sub_results[i];
+  }
+
+  return eIcicleError::SUCCESS;
+}
+
+template <typename A, typename P>
+eIcicleError
+cpu_main(const scalar_t* scalars, const A* bases, int msm_size, const MSMConfig& config, P* results, int nof_workers)
+{
+  int c = config.c;
+  if (c < 1) { c = Msm<A, P>::get_optimal_c(msm_size, config.precompute_factor); }
+
+  auto msm = Msm<A, P>{config, c, nof_workers};
 
   for (int i = 0; i < config.batch_size; i++) {
     int batch_start_idx = msm_size * i;