This PR implements a workaround to address inefficiencies in the MSM manager when handling a large number of workers. The current implementation struggles to efficiently manage and synchronize many workers, leading to performance bottlenecks.

Also note that small MSMs (logn<10/12) are really slow, and it seems the design does not account for them. 

Workaround Details:
	•	The workload is split into smaller tasks, which are assigned to thread groups and executed in parallel.
	•	Results from these smaller tasks are combined to produce the final output (in some workdivisions).
	•	Note that this is wasting threads for managers.

Future Plans:
	•	Once the MSM manager is updated to handle large numbers of workers efficiently, this workaround can be deprecated.



Seems that This is very significant for AMD, quite significant for INTEL and for M1 better in larger cases.
This feature is auto enabled but can be configured by users by specifying: threadgroup-size and total-nof-threads (via MSM config extension).

Intel i9 13900k (32 CPUs), BN254, precompute=1, batch=1 **FASTER**

- logn=14: 29ms --> 20ms
- logn=16: 82m s--> 42ms
- logn=18: 277ms --> 132ms
- logn=20: 1023ms --> 410ms (2.5X)

Intel i9 12900k (24 CPUs), BN254, precompute=1, batch=1 **SLOWER**

- logn=14: 47ms --> 76ms
- logn=16: 135ms--> 197ms
- logn=18: 490ms --> 597ms
- logn=20: 2010ms --> 2160ms (slower)

M1 pro, BN254 (10 CPUs), precompute=1, batch=1 **FASTER**

- logn=14: 34ms --> 36ms
- logn=16: 111ms--> 143ms
- logn=18: 469ms --> 387ms
- logn=20: 2091ms --> 1257ms (1.66X)

AMD EPYC 9R14 (32 CPUs) BN254, precompute=1, batch=1 **FASTER**

- logn=14: 86ms --> 25ms 
- logn=16: 265ms--> 76ms
- logn=18: 888ms --> 189ms
- logn=20: 3413ms --> 636ms (5.3X)